moved getUser() method to UserDao class, also removed Domain class
